3: Remove User
Task-Url:
https://github.com/tecnico-softeng-distsys-2015/A_15_03_17-project/issues/issue/3

Praticamente tudo feito, falta refactorizar a função loginUser.

+++ bubbledocs-appserver/src/main/java/pt/tecnico/bubbledocs/service/DeleteUser.java
@@ -2,6 +2,8 @@ package pt.tecnico.bubbledocs.service;
 
 import pt.tecnico.bubbledocs.domain.Bubbledocs;
 import pt.tecnico.bubbledocs.exceptions.BubbleDocsException;
+import pt.tecnico.bubbledocs.exceptions.UnauthorizedOperationException;
+import pt.tecnico.bubbledocs.exceptions.UnknownBubbleDocsUserException;
 import pt.tecnico.bubbledocs.exceptions.UserNotInSessionException;
 
 
@@ -22,20 +24,27 @@ public class DeleteUser extends BubbleDocsService {
     /* Verifies if give token is equal to root's token.
      * Only root can delete other users. */
     @Override
-    protected void accessControl(){
+    protected void accessControl() throws BubbleDocsException{
     	Bubbledocs bubble = Bubbledocs.getInstance();
     	String rootToken = bubble.getUserInSessionToken(root);
-    	if(!rootToken.equals(token))
+    	//User not in session cant execute the command.
+    	if(token == null || bubble.getUserFromSession(token) == null ||rootToken == null)
     		throw new UserNotInSessionException();
+    	//If the user tokens isnt root's token cant execute command.
+		else if(!token.equals(rootToken))
+			throw new UnauthorizedOperationException();
     }
     
     @Override
     protected void dispatch() throws BubbleDocsException {
     	Bubbledocs bubble = Bubbledocs.getInstance();
     	String token = bubble.getUserInSessionToken(deleteUsername);
-    	if(token != null)
+    	if(token != null)										//If user in session remove it first.
     		bubble.removeUserFromSession(token);
-    	bubble.getUserByName(deleteUsername).delete();
+    	if(bubble.getUserByName(deleteUsername) != null)		//If user exists delete it.
+    		bubble.getUserByName(deleteUsername).delete();
+    	else													//User doesnt exist.
+    		throw new UnknownBubbleDocsUserException();
     }
 
 }
